Use newer OpenGL (core) version (4.5)
This requires a small change to the fragment shader, which must now
declare an out variable, instead of using gl_FragColor.

Also, the attribute keywords seems to have been replaced with the in
keyword. And the varying keyword is replaced by out (from the vertex
shader) and in (to the fragment shader).